requirements.txt
torch>=1.8.1
torchvision>=0.9.1
timm==0.3.2
numpy
Pillow
matplotlib
opencv-python
tensorboard
wandb
submitit
segment-anything
